# openapi/media.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/projects/{id}/image-generation-size-defaults": {
      "put": {
        "tags": [
          "project-image-settings-controller"
        ],
        "operationId": "updateImageGenerationSizeDefaults",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateImageGenerationSizeDefaultsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectImageSettingsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{id}/image-generation-prompt-defaults": {
      "put": {
        "tags": [
          "project-image-settings-controller"
        ],
        "operationId": "updateImageGenerationPromptDefaults",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateImageGenerationPromptDefaultsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectImageSettingsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{id}/image-content-filter-settings": {
      "put": {
        "tags": [
          "project-image-settings-controller"
        ],
        "operationId": "updateImageContentFilterSettings",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateImageContentFilterSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectImageSettingsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{id}/article-image-resize-default": {
      "put": {
        "tags": [
          "project-image-settings-controller"
        ],
        "operationId": "updateArticleImageResizeDefault",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArticleImageResizeDefaultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectImageSettingsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/comfyui/checkpoints/install": {
      "post": {
        "tags": [
          "comfy-ui-checkpoint-controller"
        ],
        "operationId": "install",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallComfyUiCheckpointCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/projects/{id}/image-settings": {
      "get": {
        "tags": [
          "project-image-settings-controller"
        ],
        "operationId": "get_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectImageSettingsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UpdateImageGenerationSizeDefaultsRequest": {
        "type": "object",
        "properties": {
          "defaultGeneratedImageWidth": {
            "type": "integer",
            "format": "int32",
            "maximum": 2048,
            "minimum": 64
          },
          "defaultGeneratedImageHeight": {
            "type": "integer",
            "format": "int32",
            "maximum": 2048,
            "minimum": 64
          },
          "widthMultipleOf8": {
            "type": "boolean"
          },
          "heightMultipleOf8": {
            "type": "boolean"
          }
        }
      },
      "ProjectImageSettingsResponse": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "integer",
            "format": "int64"
          },
          "imageProvider": {
            "type": "string"
          },
          "comfyuiCheckpoint": {
            "type": "string"
          },
          "defaultNegativePrompt": {
            "type": "string"
          },
          "defaultQualityPrompt": {
            "type": "string"
          },
          "defaultGeneratedImageWidth": {
            "type": "integer",
            "format": "int32"
          },
          "defaultGeneratedImageHeight": {
            "type": "integer",
            "format": "int32"
          },
          "defaultArticleImageLongEdgePx": {
            "type": "integer",
            "format": "int32"
          },
          "blockSexualContent": {
            "type": "boolean"
          },
          "blockViolentContent": {
            "type": "boolean"
          },
          "blockDiscriminatoryContent": {
            "type": "boolean"
          }
        }
      },
      "UpdateImageGenerationPromptDefaultsRequest": {
        "type": "object",
        "properties": {
          "defaultNegativePrompt": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          },
          "defaultQualityPrompt": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          }
        }
      },
      "UpdateImageContentFilterSettingsRequest": {
        "type": "object",
        "properties": {
          "blockSexualContent": {
            "type": "boolean"
          },
          "blockViolentContent": {
            "type": "boolean"
          },
          "blockDiscriminatoryContent": {
            "type": "boolean"
          }
        }
      },
      "UpdateArticleImageResizeDefaultRequest": {
        "type": "object",
        "properties": {
          "defaultArticleImageLongEdgePx": {
            "type": "integer",
            "format": "int32",
            "maximum": 4096,
            "minimum": 64
          }
        }
      },
...
      "InstallComfyUiCheckpointCommand": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "integer",
            "format": "int64"
          },
          "downloadUrl": {
            "type": "string",
            "minLength": 1
          },
          "fileName": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "downloadUrl",
          "fileName",
          "jobId"
        ]
      },
...
    }
  }
}